imgui basework

+++ EngineLuke/src/ML_ENGINE/ML_ENGINE/ML/Regression/Equation.h
@@ -5,16 +5,16 @@ class Instance;
 
 struct Equation
 {
-	float a; // Slope of Line
-	float b; // Y intercept
+	double a; // Slope of Line
+	double b; // Y intercept
 
 	Equation();
 	
-	Equation(float _a, float _b);
+	Equation(double _a, double _b);
 	
 	Equation(Vector2 _point1, Vector2 _point2);
 	
 	Equation(Instance& _attr, Instance& _class);
 
-	float Distance(const Vector2& _point);
+	double Distance(const Vector2& _point);
 };
