+++ EngineLuke/src/ML_ENGINE/ML_ENGINE/ML/Regression/LinearRegression.cpp
@@ -15,11 +15,11 @@ Equation Regression::GetBestFitLine(Instance & _attr, Instance & _class)
 	return Equation(_attr, _class);
 }
 
-float Regression::SumError(Equation& line, Instance& _attr, Instance& _class)
+double Regression::SumError(Equation& line, Instance& _attr, Instance& _class)
 {
 	int size = _attr.GetAttributeCount();
 	
-	float sum = 0;
+	double sum = 0;
 	
 	for (int i = 0; i < size; ++i)
 	{
